requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydocstyle==6.1.1
pyflakes==3.0.1
pylama==8.4.1
pytz==2022.6
snowballstemmer==2.2.0
sqlparse==0.4.3
